Working Code For App
These changes include a couple new classes. The NBAPlayer class creates an NBAPlayer object which is then used to store the data pulled from an API. The Team class creates a Team Object ready to store team info from the API if needed. Changes were made to the Controller class, specifically the button. Now anytime the button in the app is clicked, the searchPlayer button pulls data for a random player from the API and displays certain information.