Weight and Height Pull Changes
There were modifications made to the controller class. It now has an if/else statement because for some players, the API does not have a recorded height or weight. So now any time a player with no height or weight gets displayed, it lets the user know that that player does not have any information available for that field.